third_party/blink/web_tests/external/wpt/webrtc/no-media-call-expected.txt

This is a testharness.js-based test.
FAIL Can set up a basic WebRTC call with no data. assert_equals: expected (number) 1 but got (object) object "[object RTCRtpSender]"
Harness: the test ran to completion.
